// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  SUPERADMIN: 'SUPERADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const InterviewType = {
  Behavioral: 'Behavioral',
  Technical: 'Technical',
  System: 'System',
  Design: 'Design',
  Mixed: 'Mixed'
} as const

export type InterviewType = (typeof InterviewType)[keyof typeof InterviewType]


export const Level = {
  Fresher: 'Fresher',
  Junior: 'Junior',
  Mid: 'Mid',
  Senior: 'Senior'
} as const

export type Level = (typeof Level)[keyof typeof Level]